requirements.txt
beautifulsoup4==4.6.3
certifi==2018.8.24
chardet==3.0.4
idna==2.7
lxml==4.2.4
python-oembed==0.2.4
requests==2.19.1
urllib3==1.23
